+++ docs/modules/changelog.rst
@@ -4,6 +4,11 @@ Full changelog
 
 This page tracks all of the changes that have been made to ``cogsworth``. We follow the standard versioning convention of A.B.C, where C is a patch/bugfix, B is a large bugfix or new feature and A is a major new breaking change. B/C are backwards compatible but A changes may be breaking.
 
+2.0.2
+=====
+- Add some more flexbility to the ``plot_cartoon_binary`` function in terms of fontsize
+- Allow use of pynbody 2.0 with cogsworth
+
 2.0.1
 =====